updated README.md file and views settings

+++ express/views.py
@@ -1,6 +1,6 @@
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticatedOrReadOnly,IsAuthenticated
-from rest_framework.decorators import api_view,permission_classes,authentication_classes
+from rest_framework.decorators import api_view,permission_classes
 from rest_framework import status
 from express.permisions import IsOwnerOrReadOnly
 from express.serializers import LogoutSerializer, ParcelPickupSerializer, ParkSerializer, PatnerSerializer, ParcelSerializer, RegisterSerializer, UserSerializer
@@ -8,8 +8,18 @@ from express.models import ParcelPickup, Park,Partner,Parcel
 from express.emails import send_email, update_email
 from django.contrib.auth.models import User
 from drf_yasg.utils import swagger_auto_schema
+from datetime import datetime
 
 
+
+
+@api_view(['GET'])
+def index_route(request):
+    now = datetime.now()
+
+    return Response({
+        'message':f'Welcome, and the server time is {now.strftime("%H:%M:%S")}'
+    })
 
 @swagger_auto_schema(method = 'POST',request_body=ParcelPickupSerializer)
 @api_view(['GET','POST'])
@@ -75,7 +85,7 @@ def partners_list(request):
     
     elif request.method == 'POST':
         data = request.data
-        email = data['email']
+        email = data['company_email']
         serializer = PatnerSerializer(data=data)
         if serializer.is_valid():
             serializer.save(creator=request.user)
@@ -101,15 +111,22 @@ def partner_detail(request,pk):
         return Response(serializer.data,status=status.HTTP_200_OK)
     
     elif request.method == 'PUT':
-        serializer = PatnerSerializer(partner,data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data,status=status.HTTP_202_ACCEPTED)
-        return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+        if request.user == partner.creator:
+            serializer = PatnerSerializer(partner,data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data,status=status.HTTP_202_ACCEPTED)
+            return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({"message":"Permision Denied"},status=status.HTTP_401_UNAUTHORIZED)
     
     elif request.method == 'DELETE':
-        partner.delete()
-        return Response({"message":"Partner Deleted"},status=status.HTTP_204_NO_CONTENT)
+        if request.user == partner.creator:
+            partner.delete()
+            return Response({"message":"Partner Deleted"},status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response({"message":"Permision Denied"},status=status.HTTP_401_UNAUTHORIZED)
+       
 
 @swagger_auto_schema(method='POST',request_body=ParcelSerializer)
 @api_view(['GET','POST'])
@@ -128,10 +145,6 @@ def parcel_track(request):
        email = parcel['parcel']['email']
        parcel_status = parcel['parcel_status']
        riders_contact = parcel['riders_contact']
-
-       print(parcel)
-    
-    
        serializer = ParcelSerializer(data=parcel)
        if serializer.is_valid():
             serializer.save()
@@ -236,15 +249,23 @@ def parks_detail(request,pk):
         return Response(serializer.data,status=status.HTTP_200_OK)
 
     elif request.method == 'PUT':
-        data = request.data
-        serializer = ParkSerializer(park,data=data)
-        if serializer.is_valid():
-            serializer.save(owner=request.user)
-            return Response(serializer.data,status=status.HTTP_201_CREATED)
-        return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+        if request.user == park.owner:
+            data = request.data
+            serializer = ParkSerializer(park,data=data)
+            if serializer.is_valid():
+                serializer.save(owner=request.user)
+                return Response(serializer.data,status=status.HTTP_201_CREATED)
+            return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({"message":"Permision Denied"},status=status.HTTP_401_UNAUTHORIZED)
     elif request.method == 'DELETE':
-        park.delete()
-        return Response({"message":"Park Deleted Successfully"},status=status.HTTP_204_NO_CONTENT)
+        if request.user == park.owner:
+            park.delete()
+            return Response({"message":"Park Deleted Successfully"},status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response({"message":"Permision Denied"},status=status.HTTP_401_UNAUTHORIZED)
+       
+            
         
 
 @swagger_auto_schema(method='POST',request_body=RegisterSerializer)
@@ -255,8 +276,8 @@ def register_view(request):
         serializer = RegisterSerializer(data=data)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data,status=status.HTTP_201_CREATED)
-        return Response(serializer.errors)
+            return Response({data:serializer.data, "message":"Account Created Successfully"},status=status.HTTP_201_CREATED)
+        return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
 
 @swagger_auto_schema(method='POST',request_body=LogoutSerializer)
 @api_view(['POST'])
